Properly parse URLs with t.co referer
URLs that contain a t.co in the referer were not being properly parsed
in objc due to a missing anchor in the TWUValidTCOURL regex. The result
was that a URL of the form:

http://foo.com/xyz?referer=http://t.co/abcde

would be parsed as http://t.co/abcde

The java, js, and ruby implementations have the correct regex. Also
added a test case for this style of URL.